fix dock

+++ shared.library/src/main/java/estg/mtsd/bikeshare/shared/library/vo/DockVo.java
@@ -7,12 +7,12 @@ public class DockVo {
 
     private Integer id;
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 
     private String location;
 
     private Integer bikeId;
-    
+
 }
